修改个人推荐二维码功能

+++ crm-infrastructure/src/main/java/com/kge/energy/crm/external/wechat/applet/req/GetUnlimitedQrCodeReq.java
@@ -14,8 +14,7 @@ public class GetUnlimitedQrCodeReq {
     @NotBlank
     private String scene;
 
-    private Double width;
-
+    private Integer width;
 
     private String env_version;
